fix(5-1): rename reducers folder -> store
